moved colors and shield icon
