+++ packages/builder/src/components/Layout.tsx
@@ -5,7 +5,7 @@ import { useAccount } from "wagmi";
 import Footer from "common/src/components/Footer";
 import { WEB3_BAD_CHAIN_ERROR } from "../actions/web3";
 import { RootState } from "../reducers";
-import colors from "../styles/colors";
+import colors from "common/src/colors";
 import { AlertContainer } from "./base/Alert";
 import Toast from "./base/Toast";
 import Landing from "./grants/Landing";
