+++ packages/builder/src/components/base/ExitModal.tsx
@@ -2,8 +2,8 @@ import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import { formReset } from "../../actions/projectForm";
 import { slugs } from "../../routes";
-import colors from "../../styles/colors";
-import Shield from "../icons/Shield";
+import colors from "common/src/colors";
+import Shield from "common/src/icons/Shield";
 import { BaseModal } from "./BaseModal";
 import Button, { ButtonVariants } from "./Button";
