+++ packages/builder/src/components/base/ImageInput.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import PinataClient from "../../services/pinata";
-import colors from "../../styles/colors";
+import colors from "common/src/colors";
 import CloudUpload from "../icons/CloudUpload";
 import ImageCrop from "./images/ImageCrop";
 import Toast from "./Toast";
