+++ packages/builder/src/components/base/TXLoading.tsx
@@ -1,5 +1,5 @@
 import Stars from "../icons/Stars";
-import colors from "../../styles/colors";
+import colors from "common/src/colors";
 import { Status } from "../../reducers/newGrant";
 
 function statusMessage(status: Status, error: string | undefined) {
