+++ packages/builder/src/components/grants/Show.tsx
@@ -10,7 +10,7 @@ import { global } from "../../global";
 import { RootState } from "../../reducers";
 import { Status } from "../../reducers/grantsMetadata";
 import { editPath, grantsPath } from "../../routes";
-import colors from "../../styles/colors";
+import colors from "common/src/colors";
 import { getProjectImage, ImgTypes } from "../../utils/components";
 import { getProjectURIComponents } from "../../utils/utils";
 import Button, { ButtonVariants } from "../base/Button";
