+++ packages/builder/src/components/rounds/ViewApplication.tsx
@@ -11,7 +11,7 @@ import Button, { ButtonVariants } from "../base/Button";
 import { Status as ApplicationStatus } from "../../reducers/roundApplication";
 import { Status as RoundStatus } from "../../reducers/rounds";
 import { grantsPath, projectPathByID } from "../../routes";
-import colors from "../../styles/colors";
+import colors from "common/src/colors";
 import Form from "../application/Form";
 import ErrorModal from "../base/ErrorModal";
 import LoadingSpinner from "../base/LoadingSpinner";
