+++ packages/common/src/components/badges/GreenVerifiedBadge.tsx
@@ -1,11 +1,11 @@
-import colors from "../../../../builder/src/styles/colors";
-import Shield from "../../../../builder/src/components/icons/Shield";
+import colors from "../../colors";
+import Shield from "../../icons/Shield";
 
 function GreenVerifiedBadge() {
   return (
     <div className="flex rounded-full bg-teal-100 px-2 py-0.5">
       <Shield dimension={16} color={colors["green-text"]} />{" "}
-      <p className="pl-1 text-green-text text-xs font-normal">Verified</p>
+      <p className={`pl-1 text-xs font-normal`} style={{ color: colors["green-text"] }}>Verified</p>
     </div>
   );
 }
